requirements.txt
pyautogui
flatbuffers
protobuf
easyocr
pyyaml